Put explorer dashboard component inside index file
- <ExplorerDashboard> replaces <GuppyDataExplorer>
